requirements.txt
# ── Core async runtime ──────────────────────────────────────────────────────
aiohttp>=3.9.0
aiofiles>=23.2.1
websockets>=12.0
asyncio-throttle>=1.0.2

# ── Exchange connectivity ────────────────────────────────────────────────────
ccxt>=4.2.0

# ── Data & time-series ──────────────────────────────────────────────────────
pandas>=2.2.0
numpy>=1.26.0
polars>=0.20.0

# ── Technical analysis ──────────────────────────────────────────────────────
ta>=0.11.0
pandas-ta>=0.3.14b
numpy>=1.26.0

# ── Machine learning ────────────────────────────────────────────────────────
torch>=2.2.0
scikit-learn>=1.4.0
lightgbm>=4.3.0
xgboost>=2.0.0
optuna>=3.5.0

# ── Reinforcement learning ───────────────────────────────────────────────────
stable-baselines3>=2.2.0
gymnasium>=0.29.0

# ── Forecasting & uncertainty ───────────────────────────────────────────────
statsmodels>=0.14.0
scipy>=1.12.0

# ── Config & validation ──────────────────────────────────────────────────────
pydantic>=2.6.0
pydantic-settings>=2.2.0
PyYAML>=6.0.1
python-dotenv>=1.0.0

# ── Database ─────────────────────────────────────────────────────────────────
asyncpg>=0.29.0           # PostgreSQL/TimescaleDB async driver
redis>=5.0.0              # Redis async client
sqlalchemy>=2.0.0

# ── Scheduling ───────────────────────────────────────────────────────────────
apscheduler>=3.10.0

# ── Monitoring & alerting ─────────────────────────────────────────────────────
python-telegram-bot>=21.0.0
httpx>=0.27.0

# ── Dashboard ────────────────────────────────────────────────────────────────
streamlit>=1.32.0
plotly>=5.19.0

# ── Logging & observability ──────────────────────────────────────────────────
structlog>=24.1.0
rich>=13.7.0

# ── Cryptography (API auth) ──────────────────────────────────────────────────
cryptography>=42.0.0

# ── Testing ──────────────────────────────────────────────────────────────────
pytest>=8.0.0
pytest-asyncio>=0.23.0
pytest-mock>=3.12.0
